pyproject.toml
[project]
name = "blog-agentic"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.119.0",
    "httpx>=0.28.1",
    "langchain>=1.0.0",
    "langchain-community>=0.4",
    "langchain-core>=1.0.0",
    "langchain-groq>=1.0.0",
    "langgraph>=1.0.0",
    "langgraph-cli[inmem]>=0.4.4",
    "pytest>=8.0.0",
    "python-dotenv>=1.0.0",
    "requests>=2.32.0",
    "streamlit>=1.41.0",
    "uvicorn>=0.38.0",
    "watchdog>=6.0.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
